名前をCanvasGroupControllerに変更
