+++ Assets/App/Scripts/Utility/Editor/UI/CanvasGroupControllerAttacher.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 631984a4b59f5491e875c7314b1fa56d
+guid: 5aa442a8cf0eb488d9834125765253b1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
