+++ Assets/App/Scripts/Utility/UI/CanvasGroupController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d659c551020404646a71f293031f2491
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
